finish split
