--- src/main/resources/isVisible
@@ -1,22 +1,0 @@
-if (!(element instanceof Element)) throw Error('DomUtil: element is not an element.');
-const style = getComputedStyle(element);
-if (style.display === 'none') return false;
-if (style.visibility !== 'visible') return false;
-if (style.opacity < 0.1) return false;
-if (element.offsetWidth + element.offsetHeight + element.getBoundingClientRect().height +
-    element.getBoundingClientRect().width === 0) {
-    return false;
-}
-const elemCenter   = {
-    x: element.getBoundingClientRect().left + element.offsetWidth / 2,
-    y: element.getBoundingClientRect().top + element.offsetHeight / 2
-};
-if (elemCenter.x < 0) return false;
-if (elemCenter.x > (document.documentElement.clientWidth || window.innerWidth)) return false;
-if (elemCenter.y < 0) return false;
-if (elemCenter.y > (document.documentElement.clientHeight || window.innerHeight)) return false;
-let pointContainer = document.elementFromPoint(elemCenter.x, elemCenter.y);
-do {
-    if (pointContainer === element) return true;
-} while (pointContainer = pointContainer.parentNode);
-return false;
