--- src/test/java/org/jdiai/entities/Header.java
@@ -1,7 +1,0 @@
-package org.jdiai.entities;
-
-import com.epam.jdi.tools.DataClass;
-
-public class Header extends DataClass<Header> {
-    public String text, visibility, tag;
-}
