--- src/test/java/org/jdiai/entities/HeaderRaw.java
@@ -1,7 +1,0 @@
-package org.jdiai.entities;
-
-import com.epam.jdi.tools.DataClass;
-
-public class HeaderRaw extends DataClass<HeaderRaw> {
-    public String innerText, className, tagName;
-}
