--- src/test/java/org/jdiai/entities/LinkHtml.java
@@ -1,5 +1,0 @@
-package org.jdiai.entities;
-
-public class LinkHtml {
-    public String tagName, src;
-}
