--- src/test/java/org/jdiai/entities/TextHtml.java
@@ -1,5 +1,0 @@
-package org.jdiai.entities;
-
-public class TextHtml {
-    public String tagName, innerText, textContent, innerHTML;
-}
