--- src/test/java/org/jdiai/entities/TextInfo.java
@@ -1,5 +1,0 @@
-package org.jdiai.entities;
-
-public class TextInfo {
-    public String tag, iText, text, iHtml;
-}
