--- src/test/java/org/jdiai/entities/UserInfo.java
@@ -1,5 +1,0 @@
-package org.jdiai.entities;
-
-public class UserInfo {
-    public String type, id, checked;
-}
