--- src/test/java/org/jdiai/tests/jsdriver/flow/TestData.java
@@ -1,18 +1,0 @@
-package org.jdiai.tests.jsdriver.flow;
-
-public class TestData {
-    public static final String SearchResultsJson =
-        "{ \"JDI SKYPE\"\"JDI Skype group with active JDI users\"\"https://join.skype.com/u2Cel0MWHkAO\" }," +
-        "{ \"JDI OWNER CONTACT\"\"Write an email directly to Framework owner and main architect Iovlev Roman\"\"mailto:roman.iovlev.jdi@gmail.com\" }," +
-        "{ \"JDI TEST SITE\"\"JDI Site for testing\"\"https://jdi-testing.github.io/jdi-light/\" }," +
-        "{ \"JDI YOUTUBE CHANNEL\"\"JDI Tutorial Official Youtube\"\"https://www.youtube.com/channel/UCck0VgwbPVgXht5h6PGCdgg/videos?view_as=subscriber\" }," +
-        "{ \"JDI DOCUMENTATION\"\"JDI Documentation\"\"https://jdi-docs.github.io/jdi-light\" }," +
-        "{ \"JDI PARTNERS GROUP - COMA QA\"\"Minsk Testing community Coma QA\"\"https://www.facebook.com/comaqa.by/\" }";
-    public static final String SearchResults =
-        "[Title: JDI SKYPE; Description: JDI Skype group with active JDI users; Link: https://join.skype.com/u2Cel0MWHkAO]," +
-        "[Title: JDI OWNER CONTACT; Description: Write an email directly to Framework owner and main architect Iovlev Roman; Link: mailto:roman.iovlev.jdi@gmail.com]," +
-        "[Title: JDI TEST SITE; Description: JDI Site for testing; Link: https://jdi-testing.github.io/jdi-light/]," +
-        "[Title: JDI YOUTUBE CHANNEL; Description: JDI Tutorial Official Youtube; Link: https://www.youtube.com/channel/UCck0VgwbPVgXht5h6PGCdgg/videos?view_as=subscriber]," +
-        "[Title: JDI DOCUMENTATION; Description: JDI Documentation; Link: https://jdi-docs.github.io/jdi-light]," +
-        "[Title: JDI PARTNERS GROUP - COMA QA; Description: Minsk Testing community Coma QA; Link: https://www.facebook.com/comaqa.by/]";
-}
