--- src/test/resources/general.xml
@@ -1,14 +1,0 @@
-<?xml version="1.0" encoding="WINDOWS-1251"?>
-<!DOCTYPE suite SYSTEM "http://testng.org/testng-1.0.dtd">
-<suite name="JSDriver Tests" parallel="methods" thread-count="8">
-    <test name="Tests">
-        <packages>
-            <package name="org.jdiai.tests.jsdriver.basics.*"/>
-            <package name="org.jdiai.tests.*"/>
-            <package name="org.jdiai.tests.jsdriver.*"/>
-        </packages>
-    </test>
-    <listeners>
-        <listener class-name="org.jdiai.testng.TestNGListener" />
-    </listeners>
-</suite>
